refactor(app_runtime): Migration handler を migration.rs に抽出 (SPEC-2077 Phase D)
- crates/gwt/src/app_runtime/migration.rs を新規作成し、 SPEC-1934 migration handler を移行: start_migration_events / skip_migration_events / handle_migration_done / handle_migration_error / quit_migration_events
- mod.rs に mod migration; 宣言を追加
- god module 解体 phase chain Phase D、 dispatch 経路 0 変更で blast radius 最小

Verification:
- cargo test -p gwt-core -p gwt: 全 pass
- cargo clippy -- -D warnings: 0 warnings
- cargo fmt --check: 差分なし
- mod.rs 6982→6865 行 (Phase D で 117 行減、 累計 580 行減)
- migration.rs 新 138 行 (SPEC-1934 migration handler 単一責務)
